backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
langchain==0.3.11
langchain-google-genai==2.0.7
langchain-community==0.3.11
langgraph==0.2.58
pydantic==2.10.2
redis==5.2.0
rq==2.0.0
pdfplumber==0.11.4
python-dotenv==1.0.1
sse-starlette==2.1.3
python-multipart==0.0.17
aiofiles==24.1.0
